fix issues with ruff
