+++ example/tasks/app.py
@@ -1,0 +1,4 @@
+"""Module for the celery app."""
+
+from parallel import *  # noqa: F403
+from serial import *  # noqa: F403
